requirements.txt
requests
feedparser
Pillow
httplib2
google-api-python-client
oauth2client
newsapi-python
